Added a link of research paper based on Fuzzing
Paper is about,
Comparasing of fuzzing with other vulnerability discovery solutions, and then introduction of the concepts and key challenges of fuzzing. 
Introduction of the state-of-the-art coverage based fuzzing, which have made great process in recent years. 
At last some techniques integrated with fuzzing, the applications and possible new trends of fuzzing.